+++ src/enum_type/EnumType.js
@@ -20,4 +20,27 @@ const ENV = {
     dev: 'dev'
 }
 
-export { MethodRequest, TransportType, ENV }
+const Codecs = {
+    OPUS: 'opus',
+    SPEEX: 'speex',
+    PCMU: 'PCMU',
+    PCMA: 'PCMA',
+    GSM: 'GSM',
+    G722: 'G722',
+    ILBC: 'iLBC',
+    ISAC: 'iSAC',
+    L16: 'L16'
+}
+
+const GraphRoute = {
+    // call
+    CallLog: 'call/history',
+    Record: 'call/recording',
+    // customer
+    Contact: 'contact',
+    AddContact: 'contact',
+    UpdateContact: 'contact',
+    DeleteContact: 'contact'
+}
+
+export { MethodRequest, TransportType, ENV, Codecs, GraphRoute }